move template data and pt report fns into correct folder
